requirements.txt
confluent-kafka
pymongo
python-dotenv
